[Refact] 함수 이름 변경
함수 역할에 맞추어 이름 변경
