+++ web/src/containers/PostDetailPage/PostSideBox/PostContent/PostText/Content/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { ContentWrapper } from './styles';
 import StyledLink from '../../../../../../components/StyledLink';
 import TimePassed from '../../../../../../components/TimePassed';
-import { parseText } from '../../../../../../lib';
+import { convertPlainTextToLinkedText } from '../../../../../../lib';
 
 function Content({ post }) {
   const { writer, content, updatedAt } = post;
@@ -19,7 +19,7 @@ function Content({ post }) {
             {writer.username}
           </StyledLink>
         </h3>
-        <article>{parseText(content)}</article>
+        <article>{convertPlainTextToLinkedText(content)}</article>
       </main>
       <TimePassed updatedAt={updatedAt} />
     </ContentWrapper>
